feat(gatsby-core-utils): improve fetch-remote-file
