+++ packages/gatsby-core-utils/src/__tests__/fetch-remote-file.js
@@ -233,24 +233,6 @@ const server = setupServer(
   )
 )
 
-function getFetchInWorkerContext(workerId) {
-  let fetchRemoteInstance
-  jest.isolateModules(() => {
-    const send = process.send
-    process.env.GATSBY_WORKER_ID = workerId
-    process.send = jest.fn()
-    process.env.GATSBY_WORKER_MODULE_PATH = `123`
-
-    fetchRemoteInstance = require(`../fetch-remote-file`).fetchRemoteFile
-
-    delete process.env.GATSBY_WORKER_MODULE_PATH
-    delete process.env.GATSBY_WORKER_ID
-    process.send = send
-  })
-
-  return fetchRemoteInstance
-}
-
 async function createMockCache() {
   const tmpDir = fs.mkdtempSync(
     path.join(os.tmpdir(), `gatsby-source-filesystem-`)
@@ -274,6 +256,7 @@ describe(`fetch-remote-file`, () => {
     // Establish requests interception layer before all tests.
     server.listen()
   })
+
   afterAll(() => {
     if (cache) {
       try {
@@ -378,316 +361,6 @@ describe(`fetch-remote-file`, () => {
     expect(filePath).toBe(cachedFilePath)
     expect(path.basename(filePath)).toBe(`logo.svg`)
     expect(gotStream).toBeCalledTimes(1)
-  })
-
-  it(`only writes the file once when multiple workers fetch at the same time`, async () => {
-    // we don't want to wait for polling to finish
-    jest.useFakeTimers()
-    jest.runAllTimers()
-
-    const cacheInternals = new Map()
-    const workerCache = {
-      get(key) {
-        return Promise.resolve(cacheInternals.get(key))
-      },
-      set(key, value) {
-        return Promise.resolve(cacheInternals.set(key, value))
-      },
-      directory: cache.directory,
-    }
-
-    const fetchRemoteFileInstanceOne = getFetchInWorkerContext(`1`)
-    const fetchRemoteFileInstanceTwo = getFetchInWorkerContext(`2`)
-
-    const requests = [
-      fetchRemoteFileInstanceOne({
-        url: `http://external.com/logo.svg`,
-        cache: workerCache,
-      }),
-      fetchRemoteFileInstanceTwo({
-        url: `http://external.com/logo.svg`,
-        cache: workerCache,
-      }),
-    ]
-
-    // reverse order as last writer wins
-    await requests[1]
-    jest.runAllTimers()
-    await requests[0]
-
-    // we still expect 2 fetches because cache can't save fast enough
-    expect(gotStream).toBeCalledTimes(2)
-    expect(fsMove).toBeCalledTimes(1)
-  })
-
-  it(`it clears the mutex cache when new build id is present`, async () => {
-    // we don't want to wait for polling to finish
-    jest.useFakeTimers()
-    jest.runAllTimers()
-
-    const cacheInternals = new Map()
-    const workerCache = {
-      get(key) {
-        return Promise.resolve(cacheInternals.get(key))
-      },
-      set(key, value) {
-        return Promise.resolve(cacheInternals.set(key, value))
-      },
-      directory: cache.directory,
-    }
-
-    const fetchRemoteFileInstanceOne = getFetchInWorkerContext(`1`)
-    const fetchRemoteFileInstanceTwo = getFetchInWorkerContext(`2`)
-
-    global.__GATSBY = { buildId: `1` }
-    let requests = [
-      fetchRemoteFileInstanceOne({
-        url: `http://external.com/logo.svg`,
-        cache: workerCache,
-      }),
-      fetchRemoteFileInstanceTwo({
-        url: `http://external.com/logo.svg`,
-        cache: workerCache,
-      }),
-    ]
-
-    // reverse order as last writer wins
-    await requests[1]
-    jest.runAllTimers()
-    await requests[0]
-    jest.runAllTimers()
-
-    global.__GATSBY = { buildId: `2` }
-    requests = [
-      fetchRemoteFileInstanceOne({
-        url: `http://external.com/logo.svg`,
-        cache: workerCache,
-      }),
-      fetchRemoteFileInstanceTwo({
-        url: `http://external.com/logo.svg`,
-        cache: workerCache,
-      }),
-    ]
-
-    // reverse order as last writer wins
-    await requests[1]
-    jest.runAllTimers()
-    await requests[0]
-
-    // we still expect 4 fetches because cache can't save fast enough
-    expect(gotStream).toBeCalledTimes(4)
-    expect(fsMove).toBeCalledTimes(2)
-  })
-
-  it(`handles 304 responses correctly in different builds`, async () => {
-    const cacheInternals = new Map()
-    const workerCache = {
-      get(key) {
-        return Promise.resolve(cacheInternals.get(key))
-      },
-      set(key, value) {
-        return Promise.resolve(cacheInternals.set(key, value))
-      },
-      directory: cache.directory,
-    }
-
-    global.__GATSBY = { buildId: `1` }
-    const filePath = await fetchRemoteFile({
-      url: `http://external.com/dog-304.jpg`,
-      cache: workerCache,
-    })
-
-    global.__GATSBY = { buildId: `2` }
-    const filePathCached = await fetchRemoteFile({
-      url: `http://external.com/dog-304.jpg`,
-      cache: workerCache,
-    })
-
-    expect(filePathCached).toBe(filePath)
-    expect(fsMove).toBeCalledTimes(1)
-    expect(gotStream).toBeCalledTimes(2)
-  })
-
-  it(`doesn't keep lock when file download failed`, async () => {
-    const cacheInternals = new Map()
-    const workerCache = {
-      get(key) {
-        return Promise.resolve(cacheInternals.get(key))
-      },
-      set(key, value) {
-        return Promise.resolve(cacheInternals.set(key, value))
-      },
-      directory: cache.directory,
-    }
-
-    const fetchRemoteFileInstanceOne = getFetchInWorkerContext(`1`)
-    const fetchRemoteFileInstanceTwo = getFetchInWorkerContext(`2`)
-
-    await expect(
-      fetchRemoteFileInstanceOne({
-        url: `http://external.com/500.jpg`,
-        cache: workerCache,
-      })
-    ).rejects.toThrow()
-
-    await expect(
-      fetchRemoteFileInstanceTwo({
-        url: `http://external.com/500.jpg`,
-        cache: workerCache,
-      })
-    ).rejects.toThrow()
-
-    expect(gotStream).toBeCalledTimes(3)
-    expect(fsMove).toBeCalledTimes(0)
-  })
-
-  it(`downloading a file in main process after downloading it in worker`, async () => {
-    // we don't want to wait for polling to finish
-    jest.useFakeTimers()
-    jest.runAllTimers()
-
-    const cacheInternals = new Map()
-    const workerCache = {
-      get(key) {
-        return Promise.resolve(cacheInternals.get(key))
-      },
-      set(key, value) {
-        return Promise.resolve(cacheInternals.set(key, value))
-      },
-      directory: cache.directory,
-    }
-
-    const fetchRemoteFileInstanceOne = getFetchInWorkerContext(`1`)
-
-    const resultFromWorker = await fetchRemoteFileInstanceOne({
-      url: `http://external.com/logo.svg`,
-      cache: workerCache,
-    })
-
-    jest.runAllTimers()
-
-    const resultFromMain = await fetchRemoteFile({
-      url: `http://external.com/logo.svg`,
-      cache: workerCache,
-    })
-
-    expect(resultFromWorker).not.toBeUndefined()
-    expect(resultFromMain).not.toBeUndefined()
-
-    jest.useRealTimers()
-
-    expect(gotStream).toBeCalledTimes(1)
-    expect(fsMove).toBeCalledTimes(1)
-  })
-
-  it(`downloading a file in worker process after downloading it in main`, async () => {
-    // we don't want to wait for polling to finish
-    jest.useFakeTimers()
-    jest.runAllTimers()
-
-    const cacheInternals = new Map()
-    const workerCache = {
-      get(key) {
-        return Promise.resolve(cacheInternals.get(key))
-      },
-      set(key, value) {
-        return Promise.resolve(cacheInternals.set(key, value))
-      },
-      directory: cache.directory,
-    }
-
-    const fetchRemoteFileInstanceOne = getFetchInWorkerContext(`1`)
-
-    const resultFromMain = await fetchRemoteFile({
-      url: `http://external.com/logo.svg`,
-      cache: workerCache,
-    })
-
-    jest.runAllTimers()
-
-    const resultFromWorker = await fetchRemoteFileInstanceOne({
-      url: `http://external.com/logo.svg`,
-      cache: workerCache,
-    })
-
-    jest.runAllTimers()
-    jest.useRealTimers()
-
-    expect(resultFromWorker).not.toBeUndefined()
-    expect(resultFromMain).not.toBeUndefined()
-    expect(gotStream).toBeCalledTimes(1)
-    expect(fsMove).toBeCalledTimes(1)
-  })
-
-  it(`downloading a file in worker process after downloading it in another worker`, async () => {
-    // we don't want to wait for polling to finish
-    jest.useFakeTimers()
-    jest.runAllTimers()
-
-    const cacheInternals = new Map()
-    const workerCache = {
-      get(key) {
-        return Promise.resolve(cacheInternals.get(key))
-      },
-      set(key, value) {
-        return Promise.resolve(cacheInternals.set(key, value))
-      },
-      directory: cache.directory,
-    }
-
-    const fetchRemoteFileInstanceOne = getFetchInWorkerContext(`1`)
-    const fetchRemoteFileInstanceTwo = getFetchInWorkerContext(`2`)
-
-    const resultFromWorker1 = await fetchRemoteFileInstanceOne({
-      url: `http://external.com/logo.svg`,
-      cache: workerCache,
-    })
-    jest.runAllTimers()
-
-    const resultFromWorker2 = await fetchRemoteFileInstanceTwo({
-      url: `http://external.com/logo.svg`,
-      cache: workerCache,
-    })
-
-    jest.runAllTimers()
-    jest.useRealTimers()
-
-    expect(resultFromWorker1).not.toBeUndefined()
-    expect(resultFromWorker2).not.toBeUndefined()
-    expect(gotStream).toBeCalledTimes(1)
-    expect(fsMove).toBeCalledTimes(1)
-  })
-
-  it(`handles 304 responses correctly in different builds and workers`, async () => {
-    const cacheInternals = new Map()
-    const workerCache = {
-      get(key) {
-        return Promise.resolve(cacheInternals.get(key))
-      },
-      set(key, value) {
-        return Promise.resolve(cacheInternals.set(key, value))
-      },
-      directory: cache.directory,
-    }
-
-    const fetchRemoteFileInstanceOne = getFetchInWorkerContext(`1`)
-    const fetchRemoteFileInstanceTwo = getFetchInWorkerContext(`2`)
-
-    global.__GATSBY = { buildId: `1` }
-    const filePath = await fetchRemoteFileInstanceOne({
-      url: `http://external.com/dog-304.jpg`,
-      cache: workerCache,
-    })
-
-    global.__GATSBY = { buildId: `2` }
-    const filePathCached = await fetchRemoteFileInstanceTwo({
-      url: `http://external.com/dog-304.jpg`,
-      cache: workerCache,
-    })
-
-    expect(filePathCached).toBe(filePath)
-    expect(fsMove).toBeCalledTimes(1)
-    expect(gotStream).toBeCalledTimes(2)
   })
 
   it(`fails when 404 is triggered`, async () => {
